refactor(workstation): dedupe project Linear tab renderers
The project-linear-projects and project-linear-work-items renderers
replicated the same LinearProjectsPage wiring (host-context callbacks,
breadcrumb/surface derivation) and differed only in the default Linear
surface. Extract the shared logic into LinearProjectsRenderer, taking
the default surface as a prop, so both files become thin wrappers.

Also extract the identical org-scope derivation shared by the
project-dashboard and project-work-items renderers into a small
getProjectOrgScopeProps helper. Behavior is unchanged.

Pre-commit hook ran. Total eslint: 0, total circular: 0
