+++ src/modules/WorkStation/TabContent/renderers/projectDashboard.tsx
@@ -9,15 +9,12 @@
 import React, { memo } from "react";
 import { useTranslation } from "react-i18next";
 
-import {
-  getProjectManagerBreadcrumbSegments,
-  getTabDataString,
-} from "@src/modules/ProjectManager/ProjectManagerLayout/components/projectManagerRouterUtils";
+import { getProjectManagerBreadcrumbSegments } from "@src/modules/ProjectManager/ProjectManagerLayout/components/projectManagerRouterUtils";
 import { useProjectHostContext } from "@src/modules/ProjectManager/ProjectManagerLayout/context/projectHostContext";
 import ProjectsPage from "@src/modules/ProjectManager/Projects";
-import { STORY_ORG_SCOPE } from "@src/store/workstation/tabs";
 
 import type { UnifiedTabContentProps } from "../types";
+import { getProjectOrgScopeProps } from "./projectOrgScope";
 
 const ProjectDashboardTabRenderer: React.FC<UnifiedTabContentProps> = memo(
   ({ tab }) => {
@@ -26,13 +23,7 @@ const ProjectDashboardTabRenderer: React.FC<UnifiedTabContentProps> = memo(
       useProjectHostContext();
 
     const breadcrumbSegments = getProjectManagerBreadcrumbSegments(tab, t);
-    const orgScope =
-      (tab.data.orgScope as string | undefined) ?? STORY_ORG_SCOPE.ALL;
-    const allowExternalSources = orgScope === STORY_ORG_SCOPE.ALL;
-    const scopedOrgId =
-      orgScope !== STORY_ORG_SCOPE.ALL
-        ? getTabDataString(tab, "orgId")
-        : undefined;
+    const { allowExternalSources, scopedOrgId } = getProjectOrgScopeProps(tab);
 
     return (
       <ProjectsPage
