+++ src/modules/WorkStation/TabContent/renderers/projectWorkItems.tsx
@@ -10,14 +10,11 @@ import React, { memo } from "react";
 import { useTranslation } from "react-i18next";
 
 import { ProjectWorkItemsTabContent } from "@src/modules/ProjectManager/ProjectManagerLayout/components/ProjectWorkItemsTabContent";
-import {
-  getProjectManagerBreadcrumbSegments,
-  getTabDataString,
-} from "@src/modules/ProjectManager/ProjectManagerLayout/components/projectManagerRouterUtils";
+import { getProjectManagerBreadcrumbSegments } from "@src/modules/ProjectManager/ProjectManagerLayout/components/projectManagerRouterUtils";
 import { useProjectHostContext } from "@src/modules/ProjectManager/ProjectManagerLayout/context/projectHostContext";
-import { STORY_ORG_SCOPE } from "@src/store/workstation/tabs";
 
 import type { UnifiedTabContentProps } from "../types";
+import { getProjectOrgScopeProps } from "./projectOrgScope";
 
 const ProjectWorkItemsTabRenderer: React.FC<UnifiedTabContentProps> = memo(
   ({ tab }) => {
@@ -30,13 +27,7 @@ const ProjectWorkItemsTabRenderer: React.FC<UnifiedTabContentProps> = memo(
     } = useProjectHostContext();
 
     const breadcrumbSegments = getProjectManagerBreadcrumbSegments(tab, t);
-    const orgScope =
-      (tab.data.orgScope as string | undefined) ?? STORY_ORG_SCOPE.ALL;
-    const allowExternalSources = orgScope === STORY_ORG_SCOPE.ALL;
-    const scopedOrgId =
-      orgScope !== STORY_ORG_SCOPE.ALL
-        ? getTabDataString(tab, "orgId")
-        : undefined;
+    const { allowExternalSources, scopedOrgId } = getProjectOrgScopeProps(tab);
 
     return (
       <ProjectWorkItemsTabContent
